chore: move configUtil from core to utils
