+++ packages/salesforcedx-utils-vscode/src/config/configUtil.ts
@@ -1,21 +1,18 @@
 /*
- * Copyright (c) 2020, salesforce.com, inc.
+ * Copyright (c) 2019, salesforce.com, inc.
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
 import {
+  Config,
   ConfigAggregator,
-  ConfigFile,
-  ConfigValue,
   OrgConfigProperties,
+  SfConfigProperties,
   StateAggregator
 } from '@salesforce/core';
-import * as path from 'path';
 import { ConfigAggregatorProvider } from '../providers';
-import { TelemetryService } from '../telemetry/telemetry';
-import { getRootWorkspacePath } from '../workspaces';
 
 export enum ConfigSource {
   Local,
@@ -23,68 +20,98 @@ export enum ConfigSource {
   None
 }
 
-// This class should be reworked or removed once the ConfigAggregator correctly checks
-// local as well as global configs. It's also worth noting that ConfigAggregator, according
-// to its docs checks local, global and environment and, for our purposes, environment may
-// not be viable.
-
 export class ConfigUtil {
   public static async getConfigSource(key: string): Promise<ConfigSource> {
-    let value = await ConfigUtil.getConfigValue(key, ConfigSource.Local);
-    if (!(value === null || value === undefined)) {
-      return ConfigSource.Local;
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const configSource = configAggregator.getLocation(key);
+    switch (configSource) {
+      case ConfigAggregator.Location.LOCAL:
+        return ConfigSource.Local;
+        break;
+      case ConfigAggregator.Location.GLOBAL:
+        return ConfigSource.Global;
+        break;
+      default:
+        return ConfigSource.None;
     }
-    value = await ConfigUtil.getConfigValue(key, ConfigSource.Global);
-    if (!(value === null || value === undefined)) {
-      return ConfigSource.Global;
-    }
-    return ConfigSource.None;
   }
 
-  public static async getConfigValue(
-    key: string,
-    source?: ConfigSource.Global | ConfigSource.Local
-  ): Promise<ConfigValue | undefined> {
-    if (source === undefined || source === ConfigSource.Local) {
-      try {
-        const rootPath = getRootWorkspacePath();
-        const myLocalConfig = await ConfigFile.create({
-          isGlobal: false,
-          rootFolder: path.join(rootPath, '.sfdx'),
-          filename: 'sfdx-config.json'
-        });
-        const localValue = myLocalConfig.get(key);
-        if (!(localValue === null || localValue === undefined)) {
-          return localValue;
-        }
-      } catch (err) {
-        if (err instanceof Error) {
-          TelemetryService.getInstance().sendException(
-            'get_config_value_local',
-            err.message
-          );
-        }
-        return undefined;
-      }
-    }
-    if (source === undefined || source === ConfigSource.Global) {
-      try {
-        const aggregator = await ConfigAggregator.create();
-        const globalValue = aggregator.getPropertyValue(key);
-        if (!(globalValue === null || globalValue === undefined)) {
-          return globalValue;
-        }
-      } catch (err) {
-        if (err instanceof Error) {
-          TelemetryService.getInstance().sendException(
-            'get_config_value_global',
-            err.message
-          );
-        }
-        return undefined;
-      }
-    }
-    return undefined;
+  /*
+   * The User-configured API version is set by the user, and is used to
+   * override the API version that is otherwise gotten from the authenticated
+   * Org in some cases, such as when deploying metadata.
+   */
+  public static async getUserConfiguredApiVersion(): Promise<
+    string | undefined
+  > {
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const apiVersion = configAggregator.getPropertyValue(
+      OrgConfigProperties.ORG_API_VERSION
+    );
+    return apiVersion ? String(apiVersion) : undefined;
+  }
+
+  public static async getDefaultUsernameOrAlias(): Promise<string | undefined> {
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const defaultUsernameOrAlias = configAggregator.getPropertyValue(
+      OrgConfigProperties.TARGET_ORG
+    );
+    return defaultUsernameOrAlias ? String(defaultUsernameOrAlias) : undefined;
+  }
+
+  public static async isGlobalDefaultUsername(): Promise<boolean> {
+    const configSource: ConfigSource = await ConfigUtil.getConfigSource(
+      OrgConfigProperties.TARGET_ORG
+    );
+    return configSource === ConfigSource.Global;
+  }
+
+  public static async getTemplatesDirectory(): Promise<string | undefined> {
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const templatesDirectory = configAggregator.getPropertyValue(
+      OrgConfigProperties.ORG_CUSTOM_METADATA_TEMPLATES
+    );
+    return templatesDirectory ? String(templatesDirectory) : undefined;
+  }
+
+  public static async isTelemetryDisabled(): Promise<boolean> {
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const isTelemetryDisabled = configAggregator.getPropertyValue(
+      SfConfigProperties.DISABLE_TELEMETRY
+    );
+    return isTelemetryDisabled === 'true';
+  }
+
+  public static async getDefaultDevHubUsernameOrAlias(): Promise<
+    string | undefined
+  > {
+    const configAggregator = await ConfigAggregatorProvider.getInstance().getConfigAggregator();
+    const defaultDevHubUserName = configAggregator.getPropertyValue(
+      OrgConfigProperties.TARGET_DEV_HUB
+    );
+    return defaultDevHubUserName ? String(defaultDevHubUserName) : undefined;
+  }
+
+  public static async getGlobalDefaultDevHubUsernameOrAlias(): Promise<
+    string | undefined
+  > {
+    const globalConfig = await Config.create({ isGlobal: true });
+    const defaultGlobalDevHubUserName = globalConfig.get(
+      OrgConfigProperties.TARGET_DEV_HUB
+    );
+
+    return defaultGlobalDevHubUserName
+      ? String(defaultGlobalDevHubUserName)
+      : undefined;
+  }
+
+  public static async getAllAliasesFor(username: string): Promise<string[]> {
+    const stateAggregator = await StateAggregator.getInstance();
+    // Without a call to clearInstance(), stateAggregator will not report
+    // aliases that were created in the current running process.
+    StateAggregator.clearInstance();
+    const aliases = stateAggregator.aliases.getAll(username);
+    return aliases;
   }
 
   /**
