+++ packages/salesforcedx-utils-vscode/test/unit/config/configUtil.test.ts
@@ -4,7 +4,11 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { OrgConfigProperties, StateAggregator } from '@salesforce/core';
+import {
+  ConfigAggregator,
+  OrgConfigProperties,
+  StateAggregator
+} from '@salesforce/core';
 import { expect } from 'chai';
 import { createSandbox, SinonStub, stub } from 'sinon';
 import {
@@ -13,56 +17,40 @@ import {
   ConfigUtil
 } from '../../../src';
 
-const vscodeStub = {
-  commands: stub(),
-  Disposable: stub(),
-  env: {
-    machineId: '12345534'
-  },
-  Uri: {
-    parse: stub()
-  },
-  window: {
-    showInformationMessage: stub()
-  },
-  workspace: {
-    getConfiguration: () => {
-      return {
-        get: () => true
-      };
-    },
-    onDidChangeConfiguration: stub()
-  }
-};
-
-describe('getConfigSource', () => {
+describe('ConfigUtil unit tests', () => {
   const sandbox = createSandbox();
-  let getConfigValueStub: SinonStub;
-
-  beforeEach(() => {
-    getConfigValueStub = sandbox.stub(ConfigUtil, 'getConfigValue');
-  });
 
   afterEach(() => {
     sandbox.restore();
   });
 
-  it('should return ConfigSource.Local if the key/value is in the local config', async () => {
-    getConfigValueStub.onCall(0).returns('someValue');
-    const configSource = await ConfigUtil.getConfigSource('key');
-    expect(configSource).to.be.eq(ConfigSource.Local);
-  });
-  it('should return ConfigSource.Global if the key/value is in the global config', async () => {
-    getConfigValueStub.onCall(0).returns(undefined);
-    getConfigValueStub.onCall(1).returns('someValue');
-    const configSource = await ConfigUtil.getConfigSource('key');
-    expect(configSource).to.be.eq(ConfigSource.Global);
-  });
-  it('should return ConfigSource.None if the key/value is not in the local or global config', async () => {
-    getConfigValueStub.onCall(0).returns(undefined);
-    getConfigValueStub.onCall(1).returns(undefined);
-    const configSource = await ConfigUtil.getConfigSource('key');
-    expect(configSource).to.be.eq(ConfigSource.None);
+  describe('getConfigSource', () => {
+    let getLocationStub: SinonStub;
+
+    beforeEach(() => {
+      getLocationStub = sandbox.stub();
+      sandbox.stub(ConfigAggregatorProvider, 'getInstance').returns({
+        getConfigAggregator: sandbox.stub().resolves({
+          getLocation: getLocationStub
+        })
+      });
+    });
+
+    it('should return ConfigSource.Local if the key/value is in the local config', async () => {
+      getLocationStub.returns(ConfigAggregator.Location.LOCAL);
+      const configSource = await ConfigUtil.getConfigSource('key');
+      expect(configSource).to.be.eq(ConfigSource.Local);
+    });
+    it('should return ConfigSource.Global if the key/value is in the global config', async () => {
+      getLocationStub.returns(ConfigAggregator.Location.GLOBAL);
+      const configSource = await ConfigUtil.getConfigSource('key');
+      expect(configSource).to.be.eq(ConfigSource.Global);
+    });
+    it('should return ConfigSource.None if the key/value is not in the local or global config', async () => {
+      getLocationStub.returns(undefined);
+      const configSource = await ConfigUtil.getConfigSource('key');
+      expect(configSource).to.be.eq(ConfigSource.None);
+    });
   });
 
   describe('getUsername', () => {
