+++ packages/salesforcedx-vscode-core/src/util/cliConfiguration.ts
@@ -5,10 +5,12 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { GlobalCliEnvironment } from '@salesforce/salesforcedx-utils-vscode';
+import {
+  ConfigUtil,
+  GlobalCliEnvironment
+} from '@salesforce/salesforcedx-utils-vscode';
 import { which } from 'shelljs';
 import { window } from 'vscode';
-import { ConfigUtil } from '.';
 import {
   ENV_SFDX_DISABLE_TELEMETRY,
   SFDX_CLI_DOWNLOAD_LINK
