+++ packages/salesforcedx-vscode-core/test/vscode-integration/conflict/metadataCacheService.test.ts
@@ -5,6 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { ConfigUtil } from '@salesforce/salesforcedx-utils-vscode';
 import {
   ComponentSet,
   FileProperties,
@@ -27,10 +28,9 @@ import {
   MetadataCacheService,
   PathType
 } from '../../../src/conflict/metadataCacheService';
+import { SfdxPackageDirectories } from '../../../src/sfdxProject';
 import { stubRootWorkspace } from '../util/rootWorkspace.test-util';
 import sinon = require('sinon');
-import { SfdxPackageDirectories } from '../../../src/sfdxProject';
-import { ConfigUtil } from '../../../src/util';
 
 describe('Metadata Cache', () => {
   describe('Metadata Cache Executor', () => {
